+++ src/app/containers/LinkedData/AudioObject.test.jsx
@@ -3,13 +3,13 @@ import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
 import AudioObject from './AudioObject';
 
 const props = {
-  promoBrandTitle: 'ماښامنۍ خپرونه',
-  shortSynopsis: 'د بي بي سي ورلډ سروس څخه پروګرام کول',
-  durationISO8601: 'PT29M30S',
-  embedUrl:
+  name: 'ماښامنۍ خپرونه',
+  description: 'د بي بي سي ورلډ سروس څخه پروګرام کول',
+  duration: 'PT29M30S',
+  embedURL:
     'https://polling.test.bbc.co.uk/ws/av-embeds/media/korean/externalId/id/ko?morph_env=live',
-  thumbnailImageUrl: 'https://ichef.bbci.co.uk/images/ic/1024x576/p063j1dv.jpg',
-  releaseDateTimeStamp: 1587655800000,
+  thumbnailUrl: 'https://ichef.bbci.co.uk/images/ic/1024x576/p063j1dv.jpg',
+  uploadDate: '2020-04-23T15:30:00.000Z',
 };
 
 describe('AudioObject', () => {
